+++ .release/release-notes.md
@@ -1,2 +1,3 @@
 ## Changes
-- Changes the executable command in bin of package.json
+- Update the action registry to handle error
+- Able to listen to action and release flags